refactor: move tools/private/zipapp to tools/zipapp
The `//tools/private` package is reserved for developer tools used to
maintain `rules_python`. The zipapp tools support rule execution and
belong under `//tools/` alongside tools like `launcher`.

Move `tools/private/zipapp` to `tools/zipapp`, update internal rule
attributes, adjust test imports and dependencies, update distribution
filegroups, and document the scope of `//tools/private`.
