+++ tests/tools/zipapp/BUILD.bazel
@@ -5,19 +5,19 @@ pytest_test(
     name = "zipper_test",
     srcs = ["zipper_test.py"],
     target_compatible_with = SUPPORTS_BZLMOD,
-    deps = ["//tools/private/zipapp:zipper_lib"],
+    deps = ["//tools/zipapp:zipper_lib"],
 )
 
 pytest_test(
     name = "exe_zip_maker_test",
     srcs = ["exe_zip_maker_test.py"],
     target_compatible_with = SUPPORTS_BZLMOD,
-    deps = ["//tools/private/zipapp:exe_zip_maker_lib"],
+    deps = ["//tools/zipapp:exe_zip_maker_lib"],
 )
 
 pytest_test(
     name = "zip_main_maker_test",
     srcs = ["zip_main_maker_test.py"],
     target_compatible_with = SUPPORTS_BZLMOD,
-    deps = ["//tools/private/zipapp:zip_main_maker_lib"],
+    deps = ["//tools/zipapp:zip_main_maker_lib"],
 )
