+++ tools/private/README.md
@@ -1,0 +1,9 @@
+# Developer Tools (`//tools/private`)
+
+This directory contains development-only tools used for maintaining and
+developing `rules_python` itself (such as release management, dependency
+updating, review bots, and repository maintenance scripts).
+
+Supporting tools for rules (e.g. `launcher`, `precompiler`, `zipapp`,
+`publish`, `wheelmaker`) belong as their own top-level directories under
+`//tools/`.
